move fix instructions panel to common folder from injected
