+++ src/common/components/cards/how-to-fix-card-row.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 import { CardRowProps } from '../../../common/configs/unified-result-property-configurations';
 import { NamedFC } from '../../../common/react/named-fc';
 import { CheckType } from '../../../injected/components/details-dialog';
-import { FixInstructionPanel } from '../../../injected/components/fix-instruction-panel';
+import { FixInstructionPanel } from '../fix-instruction-panel';
 import * as styles from './how-to-fix-card-row.scss';
 import { SimpleCardRow } from './simple-card-row';
